requirements.txt
gdown
onnx
onnxruntime
grpcio
grpcio-tools